# Specify some lambda parameters/return type explicitly