In the latest versions of the SDK I've noticed CS9236 being suggested on some test methods; this PR adds explicit typing information to the offending lambdas and makes CS9236 an error.

This knocks off ~3 seconds (~10%) from the compilation time of EFCore.Specification.Tests 🎉 :

```
Test commandline (in test/EFCore.Specification.Tests):

hyperfine --warmup 1 'touch EFCore.Specification.Tests.csproj && dotnet build-server shutdown && dotnet build'

BEFORE CHANGE:
Benchmark 1: touch EFCore.Specification.Tests.csproj && dotnet build-server shutdown && dotnet build
  Time (mean ± σ):     31.553 s ±  7.155 s    [User: 2.457 s, System: 0.612 s]
  Range (min … max):   26.723 s … 51.103 s    10 runs

AFTER CHANGE: 
Benchmark 1: touch EFCore.Specification.Tests.csproj && dotnet build-server shutdown && dotnet build
  Time (mean ± σ):     28.481 s ±  1.640 s    [User: 2.093 s, System: 0.591 s]
  Range (min … max):   26.138 s … 30.825 s    10 runs
```
/cc @artl93