@roji 2025-07-29
Follow-up for a case where CS9236 is being reported even when lambda parameters/return type are explicitly specified: https://github.com/dotnet/roslyn/issues/79650. The minimal repro there is based on one of the two EF tests where the diagnostic was suppressed.